refactoring: Extract DataFile common super class
