+++ lib/queuery_client/url_data_file.rb
@@ -1,9 +1,9 @@
+require 'queuery_client/data_file'
 require 'net/http'
 require 'stringio'
-require 'redshift_csv_file'
 
 module QueueryClient
-  class UrlDataFile
+  class UrlDataFile < DataFile
     def initialize(url)
       @url = url
     end
@@ -14,14 +14,6 @@ module QueueryClient
       @url.path
     end
 
-    def data_object?
-      /\.csv(?:\.|\z)/ =~ File.basename(key)
-    end
-
-    def gzipped_object?
-      File.extname(key) == '.gz'
-    end
-
     def open
       http = Net::HTTP.new(@url.host, @url.port)
       http.use_ssl = (@url.scheme.downcase == 'https')
@@ -30,18 +22,6 @@ module QueueryClient
         res.body
       }
       StringIO.new(content)
-    end
-
-    def each_row(&block)
-      f = open
-      begin
-        if gzipped_object?
-          f = Zlib::GzipReader.new(f)
-        end
-        RedshiftCsvFile.new(f).each(&block)
-      ensure
-        f.close
-      end
     end
   end
 end
